update warnings
- update img loader

+++ source/compat.ixx
@@ -31,7 +31,26 @@ export void CompatibilityWarnings()
     LPCWSTR szTitle = L"GTAIV.EFLC.FusionFix", szHeader = L"", szContent = L"";
     TASKDIALOG_BUTTON aCustomButtons[] = { { 1000, L"Close the program" }, { 1001, L"Continue" } };
 
-    if (GetModuleHandleW(L"d3dx9_43.dll") == NULL)
+    bool ual_present = false;
+    bool dsound_dll = false;
+    ModuleList dlls;
+    dlls.Enumerate(ModuleList::SearchLocation::LocalOnly);
+    for (auto& e : dlls.m_moduleList)
+    {
+        if (iequals(std::get<std::wstring>(e), L"dsound") && !IsModuleUAL(std::get<HMODULE>(e)))
+            dsound_dll = true;
+
+        if (IsModuleUAL(std::get<HMODULE>(e)))
+            ual_present = true;
+    }
+
+    if (dsound_dll)
+    {
+        szHeader = L"dsound.dll found";
+        szContent = L"dsound.dll in the game's root folder is not compatible with this fix\n\n" \
+            L"Please remove dsound.dll from the game's root folder and restart the game.";
+    }
+    else if (GetModuleHandleW(L"d3dx9_43.dll") == NULL)
     {
         szHeader = L"You are missing a DirectX 9 component.";
         szContent = L"It requires the latest version of " \
@@ -39,7 +58,7 @@ export void CompatibilityWarnings()
             L"<a href=\"https://www.microsoft.com/en-us/download/details.aspx?id=35\">https://www.microsoft.com/en-us/download/details.aspx?id=35</a>\n\n" \
             L"If this message keeps appearing, delete \n\nC:\\Windows\\System32\\D3DX9_43.dll\n\nC:\\Windows\\SysWOW64\\D3DX9_43.dll\n\n and install DirectX End-User Runtimes again.";
     }
-    else if (!IsUALPresent())
+    else if (!ual_present)
     {
         szHeader = L"You are running GTA IV The Complete Edition Fusion Fix with an incompatible version of ASI Loader";
         szContent = L"It requires the latest version of " \
@@ -70,6 +89,6 @@ export void CompatibilityWarnings()
     tdc.lpCallbackData = 0;
 
     auto hr = TaskDialogIndirect(&tdc, &nClickedBtn, NULL, &bCheckboxChecked);
-    if (nClickedBtn != aCustomButtons[1].nButtonID)
+    if (SUCCEEDED(hr) && nClickedBtn != aCustomButtons[1].nButtonID)
         TerminateProcess(GetCurrentProcess(), 0);
 }
